refactor(entity): 将 Notification 类中的 id 和 userId 字段类型从 Long 改为 Integer
- 修改了 Notification 类中的 id 和 userId 字段类型
- 优化了数据库表 er_notification 中相关字段的存储类型

+++ backend/src/main/java/cn/hmck/entity/Notification.java
@@ -11,9 +11,9 @@ import java.time.LocalDateTime;
 @TableName("er_notification")
 public class Notification implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Long id;  // 主键
+    private Integer id;  // 主键
 
-    private Long userId;  // 用户ID
+    private Integer userId;  // 用户ID
 
     private String message;  // 通知内容
